Added Heuristic, Feature Vector and QWDS to database. Fixed a problem with loading weights.

+++ src/main/java/players/rl/TurnSAR.java
@@ -8,11 +8,11 @@ import core.actions.AbstractAction;
 class TurnSAR {
     final AbstractGameState s;
     final AbstractAction a;
-    final float r;
+    final double r;
 
     final List<AbstractAction> possibleActions;
 
-    TurnSAR(AbstractGameState s, AbstractAction a, List<AbstractAction> possibleActions, float r) {
+    TurnSAR(AbstractGameState s, AbstractAction a, List<AbstractAction> possibleActions, double r) {
         this.s = s;
         this.a = a;
         this.r = r;
